+++ python/doc/bibliography.rst
@@ -1,0 +1,4 @@
+Bibliography
+============
+
+.. bibliography::
