+++ python/doc/refs.bib
@@ -1,0 +1,8 @@
+@misc{puchala2023casm,
+      title={{CASM Monte Carlo: Calculations of the thermodynamic and kinetic properties of complex multicomponent crystals}},
+      author={Brian Puchala and John C. Thomas and Anton Van der Ven},
+      year={2023},
+      eprint={2309.11761},
+      archivePrefix={arXiv},
+      primaryClass={cond-mat.mtrl-sci}
+}
