--- python/libcasm/monte/implementations/__init__.py
@@ -1,1 +1,0 @@
-"""Monte Carlo implementations"""
