+++ python/libcasm/monte/methods/__init__.py
@@ -1,8 +1,8 @@
 """Monte Carlo simulation methods"""
+from ._metropolis import (
+    metropolis_acceptance,
+)
 from ._monte_methods import (
     BasicOccupationMetropolisData,
     basic_occupation_metropolis,
-)
-from ._metropolis import (
-    metropolis_acceptance,
 )
