+++ python/libcasm/monte/models/ising_cpp/__init__.py
@@ -1,9 +1,9 @@
 """Python interface to C++ 2d square lattice Ising model"""
 from ._monte_models_ising_cpp import (
-  IsingConfiguration,
-  IsingState,
-  IsingSemiGrandCanonicalEventGenerator,
-  IsingFormationEnergy,
-  IsingParamComposition,
-  IsingSemiGrandCanonicalSystem,
+    IsingConfiguration,
+    IsingFormationEnergy,
+    IsingParamComposition,
+    IsingSemiGrandCanonicalEventGenerator,
+    IsingSemiGrandCanonicalSystem,
+    IsingState,
 )
