+++ python/tests/calculators/test_ising_cpp_custom_functions.py
@@ -1,17 +1,16 @@
-"""Test Python implemented property calculators with Python implemented Monte Carlo loop"""
+"""Test C++ implemented property calculators with Python implemented Monte Carlo loop"""
 import json
 import pathlib
 
-import numpy as np
-
 import libcasm.monte as monte
-import libcasm.monte.calculators.complete_semigrand_canonical_py as sgc
-import libcasm.monte.models.ising_py as ising
+import libcasm.monte.calculators.semigrand_canonical_ising_cpp as ising_impl
+import libcasm.monte.models.ising_cpp as ising
+import libcasm.monte.sampling as sampling
 
 
-def test_ising_complete_semigrand_canonical_py():
+def test_ising_basic_semigrand_canonical_cpp():
     # construct a SemiGrandCanonicalCalculator
-    mc_calculator = sgc.SemiGrandCanonicalCalculator(
+    mc_calculator = ising_impl.SemiGrandCanonicalCalculator(
         system=ising.IsingSemiGrandCanonicalSystem(
             formation_energy_calculator=ising.IsingFormationEnergy(
                 J=0.1,
@@ -22,13 +21,19 @@ def test_ising_complete_semigrand_canonical_py():
     )
 
     # construct sampling functions
-    sampling_functions = monte.StateSamplingFunctionMap()
+    sampling_functions = sampling.StateSamplingFunctionMap()
     for f in [
-        sgc.make_param_composition_f(mc_calculator),
-        sgc.make_formation_energy_f(mc_calculator),
-        sgc.make_potential_energy_f(mc_calculator),
+        ising_impl.custom_make_param_composition_f(mc_calculator),
+        ising_impl.custom_make_formation_energy_f(mc_calculator),
+        ising_impl.custom_make_potential_energy_f(mc_calculator),
     ]:
         sampling_functions[f.name] = f
+
+    json_sampling_functions = sampling.jsonStateSamplingFunctionMap()
+    for f in [
+        ising_impl.custom_make_configuration_json_f(mc_calculator),
+    ]:
+        json_sampling_functions[f.name] = f
 
     # construct the initial state
     shape = (25, 25)
@@ -36,10 +41,12 @@ def test_ising_complete_semigrand_canonical_py():
         configuration=ising.IsingConfiguration(
             shape=shape,
         ),
-        conditions=monte.ValueMap.from_dict({
-            "temperature": 2000.0,
-            "exchange_potential": [0.0],
-        }),
+        conditions=monte.ValueMap.from_dict(
+            {
+                "temperature": 2000.0,
+                "exchange_potential": [0.0],
+            }
+        ),
     )
 
     # set the initial occupation explicitly here (default is all +1)
@@ -50,14 +57,14 @@ def test_ising_complete_semigrand_canonical_py():
     event_generator = ising.IsingSemiGrandCanonicalEventGenerator()
 
     # completion check params
-    completion_check_params = monte.CompletionCheckParams()
+    completion_check_params = sampling.CompletionCheckParams()
     completion_check_params.cutoff_params.min_sample = 100
     completion_check_params.log_spacing = False
     completion_check_params.check_begin = 100
     completion_check_params.check_period = 10
 
     # Set requested precision
-    monte.converge(sampling_functions, completion_check_params).set_precision(
+    sampling.converge(sampling_functions, completion_check_params).set_precision(
         "potential_energy", abs=0.001
     ).set_precision("param_composition", abs=0.001)
 
@@ -71,6 +78,7 @@ def test_ising_complete_semigrand_canonical_py():
     mc_calculator.run(
         state=initial_state,
         sampling_functions=sampling_functions,
+        json_sampling_functions=json_sampling_functions,
         completion_check_params=completion_check_params,
         event_generator=event_generator,
         sample_period=1,
@@ -83,7 +91,7 @@ def test_ising_complete_semigrand_canonical_py():
 
     print(json.dumps(results.to_dict(), indent=2))
 
-    assert monte.get_n_samples(samplers) >= 100
+    assert sampling.get_n_samples(samplers) >= 100
     assert results.is_complete
 
     # equilibration check results
@@ -100,4 +108,3 @@ def test_ising_complete_semigrand_canonical_py():
     converge_results = results.convergence_check_results.individual_results
     for key, req in completion_check_params.requested_precision.items():
         assert converge_results[key].stats.calculated_precision < req.abs_precision
-
