+++ python/tests/calculators/test_ising_semigrand_canonical_mixed.py
@@ -2,17 +2,15 @@
 import json
 import pathlib
 
-import numpy as np
-
 import libcasm.monte as monte
+import libcasm.monte.calculators.semigrand_canonical_py as sgc
 import libcasm.monte.models.ising_cpp as ising
-from libcasm.monte.implementations.ising_cpp import (
-    SemiGrandCanonicalCalculator,
-)
+import libcasm.monte.sampling as sampling
 
-def test_ising_basic_semigrand_canonical_cpp():
+
+def test_ising_basic_semigrand_canonical_mixed():
     # construct a SemiGrandCanonicalCalculator
-    mc_calculator = SemiGrandCanonicalCalculator(
+    mc_calculator = sgc.SemiGrandCanonicalCalculator(
         system=ising.IsingSemiGrandCanonicalSystem(
             formation_energy_calculator=ising.IsingFormationEnergy(
                 J=0.1,
@@ -23,8 +21,13 @@ def test_ising_basic_semigrand_canonical_cpp():
     )
 
     # construct sampling functions
-    sampling_functions = mc_calculator.default_sampling_functions()
-    json_sampling_functions = mc_calculator.default_json_sampling_functions()
+    sampling_functions = sampling.StateSamplingFunctionMap()
+    for f in [
+        sgc.make_param_composition_f(mc_calculator),
+        sgc.make_formation_energy_f(mc_calculator),
+        sgc.make_potential_energy_f(mc_calculator),
+    ]:
+        sampling_functions[f.name] = f
 
     # construct the initial state
     shape = (25, 25)
@@ -32,10 +35,12 @@ def test_ising_basic_semigrand_canonical_cpp():
         configuration=ising.IsingConfiguration(
             shape=shape,
         ),
-        conditions=monte.ValueMap.from_dict({
-            "temperature": 2000.0,
-            "exchange_potential": [0.0],
-        }),
+        conditions=monte.ValueMap.from_dict(
+            {
+                "temperature": 2000.0,
+                "exchange_potential": [0.0],
+            }
+        ),
     )
 
     # set the initial occupation explicitly here (default is all +1)
@@ -46,14 +51,14 @@ def test_ising_basic_semigrand_canonical_cpp():
     event_generator = ising.IsingSemiGrandCanonicalEventGenerator()
 
     # completion check params
-    completion_check_params = monte.CompletionCheckParams()
+    completion_check_params = sampling.CompletionCheckParams()
     completion_check_params.cutoff_params.min_sample = 100
     completion_check_params.log_spacing = False
     completion_check_params.check_begin = 100
     completion_check_params.check_period = 10
 
     # Set requested precision
-    monte.converge(sampling_functions, completion_check_params).set_precision(
+    sampling.converge(sampling_functions, completion_check_params).set_precision(
         "potential_energy", abs=0.001
     ).set_precision("param_composition", abs=0.001)
 
@@ -67,7 +72,6 @@ def test_ising_basic_semigrand_canonical_cpp():
     mc_calculator.run(
         state=initial_state,
         sampling_functions=sampling_functions,
-        json_sampling_functions=json_sampling_functions,
         completion_check_params=completion_check_params,
         event_generator=event_generator,
         sample_period=1,
@@ -80,7 +84,7 @@ def test_ising_basic_semigrand_canonical_cpp():
 
     print(json.dumps(results.to_dict(), indent=2))
 
-    assert monte.get_n_samples(samplers) >= 100
+    assert sampling.get_n_samples(samplers) >= 100
     assert results.is_complete
 
     # equilibration check results
@@ -97,4 +101,3 @@ def test_ising_basic_semigrand_canonical_cpp():
     converge_results = results.convergence_check_results.individual_results
     for key, req in completion_check_params.requested_precision.items():
         assert converge_results[key].stats.calculated_precision < req.abs_precision
-
