+++ setup.py
@@ -8,13 +8,13 @@ setup(
         "libcasm.monte",
         "libcasm.monte.events",
         "libcasm.monte.calculators",
-        "libcasm.monte.calculators.complete_semigrand_canonical_py",
-        "libcasm.monte.implementations",
-        "libcasm.monte.implementations.ising_cpp",
+        "libcasm.monte.calculators.semigrand_canonical_ising_cpp",
+        "libcasm.monte.calculators.semigrand_canonical_py",
         "libcasm.monte.methods",
         "libcasm.monte.models",
         "libcasm.monte.models.ising_cpp",
         "libcasm.monte.models.ising_py",
+        "libcasm.monte.sampling",
     ],
     package_dir={"": "python"},
     cmake_install_dir="python/libcasm",
